backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
pydantic-settings==2.5.2
python-multipart==0.0.12
openai==1.51.0
markitdown[pdf]==0.1.1
pytest==8.3.3
httpx==0.27.2
